+++ astro.config.ts
@@ -18,6 +18,7 @@ import { themeConfig } from './src/.config'
 export default defineConfig({
   site: themeConfig.site.website,
   prefetch: true,
+  base: '/',
   image: {
     service: sharpImageService({ limitInputPixels: false }),
   },