+++ src/types/themeConfig.ts
@@ -20,6 +20,7 @@ export interface ThemeConfig {
   seo: ConfigSEO
   comment: Partial<ConfigComment>
   rss: ConfigRSS
+  analytics: ConfigAnalytics
 }
 
 export type UserConfig = DeepPartial<ThemeConfig>
@@ -30,6 +31,7 @@ export interface ConfigSite {
   author: string
   description: string
   website: string
+  pageSize: number
   socialLinks: { name: string, href: string }[]
   navLinks: { name: string, href: string }[]
   categoryMap: { name: string, path: string }[]
@@ -61,8 +63,13 @@ export interface ConfigRSS {
   follow?: { feedId: string, userId: string }
 }
 
+export interface ConfigAnalytics {
+  /** google analytics */
+  googleAnalyticsId: string
+}
+
 interface Colors {
-  foreground: string
+  primary: string
   background: string
 }